Added gpu acceleration
